Move validation utils into validation package
